requirements.txt
typing 
dotenv
langchain-core
langgraph
langchain_openai
colorama
langchain
langchain_chroma
langchain_community
pypdf
tavily-python
duckduckgo_search
beautifulsoup4
aiosqlite
langgraph.checkpoint.sqlite
pydantic

langchain
langchain-openai
langchain-anthropic
langgraph
langmem